Incomplete ST7789 TFT code added
ST7789 based TFTs are nice displays, but require some rework to get the
USB part working, and the ends up with a pretty useless frame rate.
Plus, XBM isn't really a proper fit for a color display like this, so
there really was no point to further pursuit it in this project.

But, this was the first time I used the display and I'm looking forward
to use it in a future project, and I learned and found a few gotchas
along the way that made the st7789.c code worthy to keep around, so here
we are.